insta-chat-data-extractor/requirements.txt
accelerate==1.8.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
bitsandbytes==0.46.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.4
dataclasses-json==0.6.7
deep-translator==1.11.4
diskcache==5.6.3
distro==1.9.0
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.5.1
greenlet==3.2.3
h11==0.16.0
hf-xet==1.1.4
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.0
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-community==0.3.25
langchain-core==0.3.65
langchain-text-splitters==0.3.8
langsmith==0.3.45
llama_cpp_python==0.3.9
MarkupSafe==3.0.2
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.5.0
mypy_extensions==1.1.0
networkx==3.3
numpy==2.3.0
openai==1.88.0
orjson==3.10.18
packaging==24.2
pillow==11.0.0
propcache==0.3.2
psutil==7.0.0
psycopg2==2.9.10
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.9.1
pydantic_core==2.33.2
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
safetensors==0.5.3
setuptools==70.2.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
sympy==1.13.3
tenacity==9.1.2
tokenizers==0.21.2
torch==2.7.1+cpu
torchaudio==2.7.1+cpu
torchvision==0.22.1+cpu
tqdm==4.67.1
transformers==4.53.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
yarl==1.20.1
zstandard==0.23.0
